# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API User Management",
    "version": "1.0.0"
  },
 "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:7000/api"
    }
  ],
  "paths": {
...
    "/users/lister": {
      "get": {
        "summary": "List all users",
        "operationId": "listUsers",
        "responses": {
          "200": {
            "description": "Successful response with user data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "data": {
                      "type": "array",
                      "items": { "type": "object" }
                    }
                  }
                }
              }
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ]
        }
      }
    },
    "/users/{firebaseUid}": {
      "put": {
        "summary": "Update a user",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "firebaseUid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "telephone": { "type": "string" },
                  "role_id": { "type": "string" },
                  "photo": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "data": { "type": "string" }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/users/filter/{roleId}": {
      "get": {
        "summary": "Filter users by role",
        "operationId": "filterByRole",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filtered users list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "data": {
                      "type": "array",
                      "items": { "type": "object" }
                    }
                  }
                }
              }
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ]
        }
      }
    },
    "/export-users": {
      "get": {
        "summary": "Export users to Excel",
        "operationId": "exportUsers",
        "responses": {
          "200": {
            "description": "Users exported successfully",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ]
        }
      }
    },
    "/export-users-pdf": {
      "get": {
        "summary": "Export users to PDF",
        "operationId": "exportUsersPDF",
        "responses": {
          "200": {
            "description": "Users exported successfully",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ]
        }
      }
    }
...
    "/referentiels/{id}": {
      "get": {
        "summary": "Récupérer un référentiel par ID",
        "operationId": "getReferentielById",
        "tags": ["Referentiel"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Référentiel trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Référentiel non trouvé"
          }
        }
      },
      "patch": {
        "summary": "Mettre à jour un référentiel",
        "operationId": "updateReferentiel",
        "tags": ["Referentiel"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "libelle": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "photo": {
                    "type": "string"
                  },
                  "competences": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "nom": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "duree_acquisition": {
                          "type": "integer"
                        },
                        "modules": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "nom": {
                                "type": "string"
                              },
                              "description": {
                                "type": "string"
                              },
                              "duree_acquisition": {
                                "type": "integer"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Référentiel mis à jour avec succès"
          },
          "404": {
            "description": "Référentiel non trouvé"
          }
        }
      },
      "delete": {
        "summary": "Supprimer un référentiel",
        "operationId": "deleteReferentiel",
        "tags": ["Referentiel"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Référentiel supprimé avec succès"
          },
          "404": {
            "description": "Référentiel non trouvé"
          }
        }
      }
    },
...
  },
...
  }